+++ src/Illuminate/Cache/Concerns/HasPrefix.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Illuminate\Cache\Concerns;
+
+trait HasPrefix
+{
+    /**
+     * A string that should be prepended to keys.
+     *
+     * @var string
+     */
+    protected $prefix;
+
+    /**
+     * Get the cache key prefix.
+     *
+     * @return string
+     */
+    public function getPrefix()
+    {
+        return $this->prefix;
+    }
+
+    /**
+     * Set the cache key prefix.
+     *
+     * @param  string  $prefix
+     * @return void
+     */
+    public function setPrefix($prefix)
+    {
+        $this->prefix = $prefix;
+    }
+}
